refactor: reducer を別ファイルにした
