+++ src/App.tsx
@@ -1,146 +1,9 @@
-import produce from '/web_modules/immer.js'
-import { useReducer, useEffect } from '/web_modules/preact/hooks.js'
+import { useEffect } from '/web_modules/preact/hooks.js'
+import { useRootReducer } from './reducer.js'
 import OpeningView from './OpeningView.js'
 import GameView from './GameView.js'
 import CanvasView from './CanvasView.js'
 import ResultView from './ResultView.js'
-
-type State = {
-  viewMode: 'opening' | 'game' | 'canvas' | 'result'
-  tutorial: boolean
-
-  questionState: 'drawing' | 'passed' | 'correct'
-  passCount: number
-  correctCount: number
-
-  loading: boolean
-  questions: {
-    mainText: string
-    subText: string
-  }[]
-  tutorialQuestions: {
-    mainText: string
-    subText: string
-  }[]
-}
-
-type Action =
-  | {
-      type: 'showResult'
-    }
-  | {
-      type: 'closeResult'
-    }
-  | {
-      type: 'startDrawing'
-    }
-  | {
-      type: 'startGame'
-    }
-  | {
-      type: 'startTutorial'
-    }
-  | {
-      type: 'passQuestion'
-    }
-  | {
-      type: 'correctQuestion'
-    }
-  | {
-      type: 'goToNextQuestion'
-    }
-  | {
-      type: 'resetGame'
-    }
-  | {
-      type: 'loadStart.sheetValues'
-    }
-  | {
-      type: 'loadEnd.sheetValues'
-      payload: {
-        sheetValues: (string | boolean)[][]
-      }
-    }
-
-const reducer = produce((state: State, action: Action) => {
-  switch (action.type) {
-    case 'showResult':
-      state.viewMode = 'result'
-      break
-
-    case 'closeResult':
-      state.viewMode = 'game'
-      break
-
-    case 'startDrawing':
-      state.viewMode = 'canvas'
-      break
-
-    case 'startGame':
-      state.tutorial = false
-      state.viewMode = 'game'
-      break
-
-    case 'startTutorial':
-      state.tutorial = true
-      state.viewMode = 'game'
-      break
-
-    case 'passQuestion':
-      state.viewMode = 'game'
-      state.questionState = 'passed'
-      state.passCount += 1
-      break
-
-    case 'correctQuestion':
-      state.viewMode = 'game'
-      state.questionState = 'correct'
-      state.correctCount += 1
-      break
-
-    case 'goToNextQuestion':
-      state.questionState = 'drawing'
-      break
-
-    case 'resetGame':
-      state.viewMode = 'opening'
-      state.questionState = 'drawing'
-      state.passCount = 0
-      state.correctCount = 0
-      break
-
-    case 'loadStart.sheetValues':
-      state.loading = true
-      break
-
-    case 'loadEnd.sheetValues':
-      const { sheetValues } = action.payload
-
-      state.loading = false
-
-      const parsed = sheetValues
-        .slice(1)
-        .map(([mainText, subText, forTutorial, disabled]) => {
-          if (disabled || (!mainText && !subText)) {
-            return null
-          }
-
-          return {
-            mainText: mainText as string,
-            subText: subText as string,
-            forTutorial: Boolean(forTutorial),
-          }
-        })
-        .filter(nonNull)
-
-      state.questions = shuffle(parsed.filter(q => !q.forTutorial))
-      state.tutorialQuestions = shuffle(parsed.filter(q => q.forTutorial))
-      break
-
-    default:
-      const _: never = action
-  }
-})
 
 export default function App() {
   const [
@@ -155,16 +18,7 @@ export default function App() {
       tutorialQuestions,
     },
     dispatch,
-  ] = useReducer<State, Action>(reducer, {
-    viewMode: 'opening',
-    tutorial: false,
-    questionState: 'drawing',
-    passCount: 0,
-    correctCount: 0,
-    loading: false,
-    questions: [],
-    tutorialQuestions: [],
-  })
+  ] = useRootReducer()
 
   const showResult = () => dispatch({ type: 'showResult' })
   const closeResult = () => dispatch({ type: 'closeResult' })
@@ -265,24 +119,6 @@ function fetchSheetValues() {
       setTimeout(() => resolve(p), 1000)
     }
   })
-}
-
-// https://medium.com/@nitinpatel_20236/how-to-shuffle-correctly-shuffle-an-array-in-javascript-15ea3f84bfb
-function shuffle<T>(array: T[]): T[] {
-  const target = [...array]
-
-  for (let i = target.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * i)
-    const temp = target[i]
-    target[i] = target[j]
-    target[j] = temp
-  }
-
-  return target
-}
-
-function nonNull<T>(value: T | null | undefined): value is T {
-  return value !== null && value !== undefined
 }
 
 // FIXME ベタがきデータ
